add multi-call kairos binary; move sub-tool CLIs into importable packages
Introduce cmd/kairos, a single binary that dispatches to immucore,
kairos-agent, or the kcrypt-discovery-challenger based on its argv[0]
(the name it is invoked as, so a symlink named "immucore" runs the
immucore code path). The explicit form "kairos <sub-tool> [args]"
also works. Same pattern as busybox.

To make that possible, each sub-tool's CLI wiring moves out of its
own main.go into an importable package:

  immucore/pkg/cmd            (NewApp, Run)
  agent/pkg/cmd               (Run)
  kcrypt/discovery            (Run; was kcrypt/cmd/discovery/main.go)
  kcrypt/challenger           (Run; was kcrypt/main.go)

Standalone builds of each sub-tool still work; each main.go is now a
thin stub that calls into the package. The kcrypt discovery client's
"client" helper moved from kcrypt/cmd/discovery/client to
kcrypt/discovery/client along with it; kcrypt/tests updated to match.

The kcrypt challenger server (kubernetes controller manager) builds
as its own binary at cmd/kcrypt-challenger. It stays out of the
multi-call kairos binary because it pulls in controller-runtime and
client-go, which have no place on a device.

register_agent.go is gated with //go:build !initramfs so
"go build -tags initramfs ./cmd/kairos" drops the kairos-agent
import and its transitive deps, yielding a smaller binary suitable
for shipping in an initramfs (where kairos-agent never runs).

Sizes, stripped with -trimpath -ldflags="-s -w":

  kairos (full)            33M   argv[0] names: immucore, kairos-agent,
                                 kcrypt-discovery-challenger
  kairos (-tags initramfs) 25M   kairos-agent not linked
  kcrypt-challenger        32M   in-cluster server, its own artifact
