+++ cmd/kairos/register_immucore.go
@@ -1,0 +1,9 @@
+package main
+
+import (
+	immucore "github.com/kairos-io/kairos/immucore/pkg/cmd"
+)
+
+func init() {
+	register("immucore", immucore.Run)
+}
