+++ cmd/kairos/register_kcrypt.go
@@ -1,0 +1,9 @@
+package main
+
+import (
+	kcrypt "github.com/kairos-io/kairos/kcrypt/discovery"
+)
+
+func init() {
+	register("kcrypt-discovery-challenger", kcrypt.Run)
+}
