+++ kcrypt/discovery/cli_test.go
@@ -1,4 +1,4 @@
-package main
+package discovery
 
 import (
 	"os"
