+++ kcrypt/discovery/client/client.go
