+++ kcrypt/discovery/client/mdns.go
